Removing legacy code and STL
I noticed that the size of object and the binary files increased significantly using STL in the code. I started removing STL from the code in most and substituting it with C standart library and my own solutions. In future, I will start removing other dependencies and writing everything myself.