Windows-specific fixes
Fixed compilation error with clipboard handling on Windows, rewritten Timer in Windows ( I used to use SDL time API for this, now I use platform-specific functions: in Windows QueryPerformanceFrequency and QueryPerformanceCounter, in Unix gettimeofday). Fixed bug with attaching texture to framebuffer when texture is not attached to framebuffer. In the future, I will create more complex system for this, but now this is not neccessary.