Upgraded editor, fixed bugs, ResourceManager
So today I have created new MessageBox class for editor, ResourcesViewerTexture to load textures to scene and select them in Inspector to attach to the material. I've upgrated FileDialog, now you can select multiple files by selecting them with holded Ctrl key. And the main feature is ResourceManager. Now it used only for textures but in future it will automate all the process of managing resources, loading and saving them in right places.